refactor states module
